Add files via upload
This is an R script that maps the trips between different cities in Namibia. There are three components:

1. importing the shape file for Namibia
2. importing the city location file
3. importing the trip data

The shape file is used to make a map of Namibia with the states outlined. 
The city locations is used to determine (1) where the cities are on the map and (2) to define the start and end of each trip
The trip data is used to determine the number of trips between cities